fix(data-warehouse): split light hooks out of heavy temporal facade
AppConfig modules (revenue_analytics, signals) import the data-import hooks during django.setup(); routing them through facade.temporal pulled settings (temporalio/dlt/pandas/posthog.schema) onto the startup path, failing the import-budget test. New facade.hooks re-exports only the light external_product_hooks surface; facade.temporal keeps the heavy ACTIVITIES/WORKFLOWS for off-setup consumers (worker bootstrap, temporal activities).
